refactor: drop dead dialog overloads + stale references
- Remove FlipcashUI's dialog(isPresented:style:...) and
  dialog<T>(item:style:...) overloads + their modifier structs.
  Zero call sites repo-wide — the app-target dialog(item:) modifier
  bypasses both. Rename the surviving Dialog+View.swift to
  ActionBuilder.swift to reflect what's left.
- Drop the unused error parameter on
  OnboardingViewModel.showAccountCreationError. Both call sites passed
  an error that the body never read.
- Update the stale staleRate comment on GiveViewModel.prepareSubmission
  to describe the rate-unavailable error the caller now surfaces.
